data_collector/requirements.txt
requests==2.10.0
websocket-client==1.2.1
psycopg2==2.9.1
PyYAML==5.4.1
